New UrlMetrics fields, and support for calculating 'missing' fields.
Added constants for all of the new fields now returned by UrlMetrics.

Added a list of fields whose values are not returned by UrlMetrics,
but which can be calculated from other fields. For example,
followed internal links can be calculated from total followed links
minus followed external links. If the neccessary fields are present
in the response, the calculated fields will automatically be added.
If you request the calculated fields, the request will actually be
for the fields that it depends upon.

[Internal tracking: TUBO-869]